DEV-11854 : 버그 수정.
